# linkso_server/openapi.yaml
openapi: 3.1.0
info:
  title: LinkSo API
  version: 1.0.0
  description: HTTP contract implemented by the LinkSo Rust server and consumed by the Flutter client.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/admin/ad-campaigns:
    post:
      tags: [admin]
      summary: Create an advertising campaign
      security: [{adminToken: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CampaignWrite'}
      responses:
        '201': {description: Campaign created}
        '401': {$ref: '#/components/responses/ApiError'}
  /api/v1/admin/ad-campaigns/{id}:
    put:
      tags: [admin]
      summary: Update an advertising campaign
      parameters: [{$ref: '#/components/parameters/Id'}]
      security: [{adminToken: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CampaignWrite'}
      responses:
        '200': {description: Updated campaign}
        '401': {$ref: '#/components/responses/ApiError'}
        '404': {$ref: '#/components/responses/ApiError'}
  /api/v1/admin/ad-campaigns/{id}/enable:
    post:
      tags: [admin]
      summary: Enable a campaign
      parameters: [{$ref: '#/components/parameters/Id'}]
      security: [{adminToken: []}]
      responses:
        '200': {description: Enabled campaign}
        '401': {$ref: '#/components/responses/ApiError'}
        '404': {$ref: '#/components/responses/ApiError'}
  /api/v1/admin/ad-campaigns/{id}/disable:
    post:
      tags: [admin]
      summary: Disable a campaign
      parameters: [{$ref: '#/components/parameters/Id'}]
      security: [{adminToken: []}]
      responses:
        '200': {description: Disabled campaign}
        '401': {$ref: '#/components/responses/ApiError'}
        '404': {$ref: '#/components/responses/ApiError'}
  /api/v1/ad-campaigns/active:
    get:
      tags: [public-flow]
      summary: Get the currently active public campaign
      responses:
        '200': {description: Active campaign}
        '204': {description: No active campaign}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: linkso_session
    bearerAuth:
      type: http
      scheme: bearer
    adminToken:
      type: apiKey
      in: header
      name: Authorization
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: {type: string, format: uuid}
...
  responses:
...
    ApiError:
      description: Structured API error with a request ID
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
...
    CampaignWrite:
      type: object
      additionalProperties: false
      required: [title, body, advertiser_url, starts_at, ends_at]
      properties:
        title: {type: string}
        body: {type: string}
        image_url: {type: [string, 'null'], format: uri}
        advertiser_url: {type: string, format: uri}
        starts_at: {type: string, format: date-time}
        ends_at: {type: string, format: date-time}
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code: {type: string}
            message: {type: string}
            request_id: {type: string}